feat: introduce a download button in the app

+++ src/utils/formatDate.js
@@ -4,7 +4,7 @@ export default function formatDate(date) {
     let year = new Intl.DateTimeFormat("en", { year: "numeric" }).format(
       formattedDate
     );
-    let month = new Intl.DateTimeFormat("en", { month: "long" }).format(
+    let month = new Intl.DateTimeFormat("en", { month: "short" }).format(
       formattedDate
     );
     return `${month} ${year}`;
